Fixes test error caused by empty pickle
Fixes the bug as discussed on Slack.

Furthermore, tests have been refactored to rely on fixtures for the
model pickle file and the config.json, instead of checking those files
into the repo. The files are cleaned up after the tests finish running.
Existing tests have been adapted to use the fixtures.