ENH add examples and col names to CardData and config.json (#45)
Users now need to pass data to init. This is being used to infer the column
names from the data (generated names for numpy arrays, column names for
DataFrames) and the first 3 samples are used to present as examples on the
Hub.

Note that the data can also be used in the future to store the dtypes of the
column.