Small refactor: figure_paths -> extra_sections
Functionally, nothing changed, this is just a renaming and a little
cleanup.

The reason is that this attribute can be used for more than just
figures, so a more generic name is a better fit. The name
"extra_sections" was chosen to mirror the "template_sections".

Apart from renaming and adjusting some comments based on that, there was
a small change in which the items of the dict are accessed, making it
more efficient, and merging to f-strings into one.