Extend user guide to include classification report
It is currently already possible to add classification reports to the
model card, but it might not be completely obvious. Therefore, the model
card user guide now includes an example of how to add the classification
report.

Some of the texts in the user guide are rewritten to (hopefully) provide
more clarity and to remove some outdated info.