Remove redundant test in CI
As discussed, the "Tests" and "Doc Tests" tasks are redundant. Since CI
is already quite slow as is, we really don't want to run all the tests
twice. Therefore, the "Doc Tests" task is removed.

Ideally, we would have one task for only doctests and one for unit tests
but AFAICT, there is no option to _only_ run doctests.