Extend tests to cover learned attributes
This is useful e.g. for clustering models that don't have a predict
method.

While adding these tests. A few estimators started to fail. Most of them
failed because of the discussed issue of numpy scalars being loaded as
0-dim arrays. The issue was fixed via explicit type casting.

However, 3 estimators are now failing and had to be added back to the
list of failing estimators. I added a comment for each one of them why
they fail and how to potentially address the issue.