Roll back test changes, use pytest tmp_path
From my understanding, pytest will (eventually) clean up the temporary
files it creates. Therefore, use the pytest tmp_path fixture instead of
the builtin tempfile module and don't explicitly clean up.

What's strange is that for other tests, this wasn't necessary, not sure
why it is here. It may have to do with how we save and load here or with
the use of zip files. Ideally, someone with access to Windows could test
it.