FEA add secure persistence (#128)
This is the first implementation of a secure persistence feature for sklearn. By
not using pickle, this feature aims at being more secure, so that untrusted
sklearn models can be securely loaded by users.

At this stage, this is still an experimental feature and should not be relied
upon to be secure. Most of sklearn is working (the known exception being Birch).
More testing is needed and feedback appreciated.

Co-authored-by: Benjamin Bossan <benjamin.bossan@gmail.com>