Avoid duplicating numpy arrays
Fixes #135

Instead of using a UUID for the name of the numpy files, use their hash.
That way, when the same array already exists, we can use it instead of
storing another copy.

Implementation

I started out with using hashlib.sha256 instead of delegating hashing to
joblib. However, this caused several problems:

- sha256 only worked for C-contiguous arrays
- the sha256 of np.array(0), np.array([0]), np.array([[0]]) etc. were
  all the same, which prevents us from restoring the arrays to their
  original shape

I did not try the other hashlib algorithms. Both drawbacks could be
worked around (store the contiguity and shape and restore them after
loading), but since joblib is already a dependency, I just went with it.

The disadvantage of joblib.hash is that it uses either md5 or sha1, both
of which are not cryptographically secure anymore. However, for this use
case, that should not be a necessity, since the hash only serves as a
cache key.