Bugfix when encountering duplicate np arrays
Previously, when a duplicate numpy array was encountered, the file was
just overwritten with the same content. Now, the persistence is skipped.
This has no functional consequences but is more efficient.

While working on this, I discovered a minor bug where trying to store an
object numpy array resulted in the creation of a broken .npy file being
left over. This is because numpy tries to write to the file until it
encounters an error and raises, but then doesn't clean up said file.
Before this bugfix, we would include that broken file in the zip
archive, although we wouldn't do anything with it. Now, no such file is
being created. A test is added to test this.