Fix bug with line breaks in markdown tables
Markdown tables shouldn't contain "\n" line breaks. Instead, they should
be replaced by "<br />". This is now done for both the hyperparameter
table and for tables added through card.add_table.

Implementation

The bugfix has to be applied after calling tabulate, as the table
entries can contain arbitrary objects, whose repr is only called by
tabulate. That also means we have to be careful as the tabulated table
string will contain valid "\n" line breaks that should be left
untouched.

Unfortunately, the hyperparameter table doesn't use card.add_table,
which is why the fix has to be applied in two places. IMO it would be
better if it used the same method under the hood but that would result
in the hyperparameter table being placed in a different section (because
of the inflexibility of the templating).

A minor naming issue I found: The method _extract_estimator_config
actually creates the hyperparameter table. I would have never guessed to
be honest.