Relax dependency of HF hub being < v0.10
That restriction was a short term solution because we thought we could
make v0.10 work with a few changes. However, that's not the case because
of a bug in HF hub model cards on Windows. This requires a bugfix
release before it would be fixed.

In the meantime, we should allow users to install v0.10, since there is
no technical reason not to use it. The only annoyance is that users will
get FutureWarnings.

For CI, we filter out corresponding FutureWarnings for the time being.
Once the HF hub bug is fixed, we can proceed with #162 and remove the
filter.