[WIP] Add dumps and loads functions
Resolves issue #180. Quoting:

It would be nice to have the equivalent of pickle.dumps / pickle.loads
or json.dumps / json.loads for skops.save / skops.load. That way, saving
and loading does not necessarily require dealing with files. It could
also facilitate testing and sending data over the wire.

Implementation

This is still very much WIP but I wanted to get some early feedback. The
idea is to reuse as much code as possible. The main changes are around
saving numpy arrays and sparse matrices, which now have to deal with a
third option.

Regarding the tests, I have modified them so that they now run once with
save/load and once with dumps/loads. This of course doubles the time
required to run those tests, so in the future we may only run one
setting (dumps and loads) and then have a handful of tests that
specifically test save/load.