[WIP] Persist: Add auditing feature
The goal of this PR is to add an auditing feature which is used while an
object is being loaded. Before the actual object is loaded, the state is
inspected and some audits are performed. (By default) if at least one
audit fails, the loading process is interrupted and the user is informed
about what went wrong.

Functionally, this does more or less the same as Adrin's work on
https://github.com/adrinjalali/skops/tree/audit only with much more code
;)

Implementation

The reason why much more code is used here is that I aimed at making the
auditing extensible. This way, it would be possible in the future for
each user to add their own auditing functions on top of the ones used in
skops if they require additional security.

Also, in contrast to Adrin's branch, this implementation does not
traverse the nodes of the state recursively. Instead, each get_instance
function is decorated by an auditing object which is responsible for
calling each audit function before the respective object (or suboject)
is being loaded.

This is better than traversing the nodes because here, we can re-use the
node traversing that's already implemented, making absolutely sure that
no node is being missed.

At the moment, the actual auditing functions are very rough and do not
actually help much with security. This is okay, as this draft PR should
serve to discuss the general implementation. Implementing the actual
auditing functions can be done later.

Furthermore, we have "false positives", e.g. certain sklearn functions
which are marked as insecure. Therefore, our current testing suite would
fail with the given auditing functions. However, I configured the tests
so as to not raise an error when encountering an insecure node, instead
just logging the results.

I will also add more unit tests (there are currently only 2 very
rudimentary ones) if and when we decide to go ahead with this
implementation.