Bugfix for _DictWithDeprecatinKeys not importable
In sklearn 1.2.0, _DictWithDeprecatinKeys will be removed. Currently,
skops will fail to import it and raise an error. In the future, we try
to import it and if that fails, not add it as a dispatch type.

Implementation

The bugfix should work. There is, however, trouble with testing it,
because we need to patch top level imports. This does not work properly
because the order of patching vs importing is wrong, resulting in the
patch being applied too late.

I have added a test to demonstrate it. If it is run in isolation, it
passes:

$ python -m pytest skops/io/tests/test_deprecation.py

If run in conjunction with the other tests, it fails:

$ python -m pytest -m skops
E           assert not True
E            +  where True = any(<generator object test_dictwithdeprecatedkeys_cannot_be_imported.<locals>.<genexpr> at 0x7fa727c2e3b0>)

This is because the other tests will trigger the import of
_DictWithDeprecatinKeys _before_ the patch is applied. This is bad,
ideally we want our tests not to affect each other.

I see these solutions:

1. remove the test and just assume that the bugfix works
2. CI matrix extended to run with sklearn<1.2 and with sklearn>=1.2
3. change imports such that the tests don't influence each other

I think it goes without saying why 1 and 2 are not very desirable.

I haven't actually tested 3. but the idea would be to make all skops
imports in the tests local instead importing on the root level. So for
example if we want to test skops.io.load, instead of having:

from skops.io import load

def test_it():
    load(...)

we should have:

@pytest.fixture
def load():
    from skops.io import load
    return load

def test_it(load):
    load(...)

If I'm not mistaken, this should resolve this issue and most similar
issues in the future (it's not a cure all solution though).

As a practical example, in skorch we handle it exactly like this. Below
is an example of testing the skorch to_tensor function. See how the
function is only imported locally, not on a module level:

https://github.com/skorch-dev/skorch/blob/cfe568b58f150730e7269b5a2a1b2bac2228dbd2/skorch/tests/test_utils.py#L15-L34

Hopefully, there is a better way than any of the suggestions, but I
couldn't come up with any.