Skip inference tests w/ using sklearn dev version
Pulling a VW here: When detecting that the sklearn version is a dev
version (like the nightly build), skip the inference tests. This is
because the inference environment will fail to build the docker image
for those sklearn versions, as they're not on the conda package index.

Just installing with pip would not be a solution, because the sklearn
nightly build lives on a different index.