Make skops.io.dump and load work with TextWrapper
(note: load already worked, only dump needed change)

The idea here is to make it possible to use dump and load with a file
wrapper, i.e. using 'with open(...)'. This makes it easier to search and
replace pickle dump and load by skops dump and load.

Implementation

To decide how to check the way to treat the file argument, I went with
hasattr(file, "write"), as this is what np.save does. I assume this is
more robust than checking the type. Unfortunately, mypy doesn't
understand this, so I had to tell it to shut up.

I didn't update the docstring or any examples because I don't think we
need to 'advertise' this usage.