Add more test cases for model card parser
I added 5 model cards from the top 10 most used models from the Hub (I
excluded cards that were too similar to one another). The tests were
rewritten so that they should now pass.

There are some limitations to the parser that results in the generated
cards not being 100% identical. Those limiations are now documented.
However, I don't believe that those limitations matter, as they make no
semantic difference but rather are stylistic or even invisible. The most
notable difference is the alignment of columns in tables.

The tests pass despite those differences because I rewrote them to
include a diff file for each model card. When the generated card is
compared to the original one, a diff is created and compared to the
checked in diff. This way, we have control over what diff we permit.

I had to exclude the folder containing the cards and diffs from the
pre-commit task "trailing-whitespace", as we need the trailing
whitespace in there.