FEAT: Alternative Card implementation (#203)
Description

The proposed model card implementation would allow to dynamically add
sections or overwrite them.

This is not a complete implementation but already covers most of the
features we already have and then some.

On top of these features, it would be possible to add more features like
creating a default Card with placeholders, just like the exisint
template, or the possibility to delete existing sections or to retrieve
the result of a certain section.

Implementation

The underlying data structure consists of a dict and a Section
dataclass.

All data is stored in a _data attribute with the type dict[str,
Section]. The dataclass hold the section contents, i.e. the section
title, the section content, and subsections, which again have the same
type. It's thus recursive data structure. Section title and dict key are
identical, which is mostly for convenience.

With this refactor, there are no separate data containers anymore for
eval results, template sections, extra sections, etc. They are all
treated the same.

IMHO, this greatly simplifies the code overall. The only complex
function that's left is the one needed to traverse the tree holding the
data, and even that is just 14 LOC.

Demo

To see how the new class can be used, take a look at the main function.
The resulting Card can be seen here:

https://huggingface.co/skops-ci/hf_hub_example-fcc0d6fe-d072-4f94-8fdb-6bf3bb917bca

* [WIP] Further align new model card design

Added a test that shows that the new card produces the same output as
the old card (except for a few non-deterministic parts). This includes
most of the idiosyncrasies of the old card we might want to change in
the future (e.g. inconsistent capitalization, use of empty lines). Some
of the more problematic behaviors of the old card class were, however,
fixed (e.g. creating an empty metrics table when there are no metrics).

The other tests have been reworked to use the new card features to make
them more precise. Often, that means that instead of having a very weak
test like "assert 'foo' in card.render()", it is now possible to select
the exact section and check that it equals the expected output.

This work is still unfinished, specifically it still lacks tests for the
card repr and for the newly added features.

* Make tests pass

Some refactoring to clean up things, rework repr, make repr tests pass.

* Add tests for new functionalities and docstrings

* Adjust tests to work with older sklearn versions

* Adjust examples to new card, more docs

Also some better type annotations.

* Continue fixing tests

* Try fixing Windows error by specifying encoding

* Adjust doctest: confusion matrix not stored in cwd

* Increase test coverage

* Try fixing test failure on Windows

* Replace old by new Card implementation

* Address reviewer comments

- Remove noise from docstring example
- Add the comma after model repr
- Add docstrings to private methods

* Add TODO notes for when Python 3.7 is dropped

* Add Hub model card template, add template arg

Users can now choose to use no template, skops template, hub template,
or their own template. Using their own template disables a lot of
prefilling (say, putting the model plot in the card) because we wouldn't
know where to put it. Users will need to call card.add for the otherwise
prefilled sections.

* Make _add_single return the Section

This can be useful, because otherwise it takes a bit of effort to
retrieve the latest section.

* Allow tables without rows to be added

It's ugly, but there is no technical reason from prohibiting the
addition of tables without rows. (Note, columns are still required).

This allows us to use TableSection for formatting the metrics, instead
of calling tabulate there directly. This is better, since we don't have
2 separate ways of creating metrics.

* Error when calling add_metric w/ invalid template

* Some amendments required after merging

Most notably, dynamic model loading in Cards was broken.

* A couple more changes:

1. Don't use Hub template for now
2. Document how to add new default templates in code
3. Nicer error message when using invalid template

* Add entry to changes.rst

* Fix small bug in generated code for loading

* Address reviewer comments

- Remove Python 3.7 compatibility code (requires #246)
- Removed empty section in docs
- Don't add empty metrics table by default
- Remove option to select with a list of str (e.g. card.select(['Foo',
  'Bar'])) is no longer possible)
- Add option to chain selections using select method (e.g.
  card.select('Foo').select('Bar'))

* Adjust getting started code in model card

* Adjust docstrings

* Allow adding default sections with custom template

So far, if users had a custom template (including no template at all),
they would lose the possibility to add some default sections:

- metrics
- model plot
- hyperparamters
- getting started code

Now we expose methods to the users to add those sections with a simple
call (no need to manually format the content and use card.add). For this
to work, users have to indicate the desired section, since we would
otherwise not know where to put the content.

During the work on this, I also cleaned up the Card tests, which became
messier and messier over time. Related tests are now all contained in a
class, which makes it a little bit easier to see if for a certain
method, all test cases have been covered.