Turn off CatBoost verbosity in tests
By default, catboost prints a giant output to stdout when fitting. This
is annoying when running the tests, which include fitting catboost.
Normally, this does not matter, as pytest captures stdout, but when
running pytest with -s (e.g. for debugging purposes or in CI), the
output is flooded.

This PR adds verbose=False when initializing the catboost estimators,
which prevents any output from being shown.