Work on supporting scikit-learn intelex
Partially solves #251

This is one part of the work required to solve the mentioned issue. The
other part will have to be added on the API inference side, once this PR
is finalized.

Description

Added the option use_intelex to hub_utils.init. It adds a new entry to
config.json which can later be used by the inference API to decide
whether to run it with intelex or not.

On top of that, if use_intelex=True, scikit-learn-intelex will be added
as a dependency to the requirements if not already there. Moreover, if a
metadata for a model card is loaded, a scikit-learn-intelex tag will be
added.