ENH Add possibility to turn section invisible
Sections of model cards have an additional flag called visible. It is by
default True but if set to False, the corresponding section, and its
subsections, are not rendered.

In contrast to deleting those section, turning them invisible will not
remove them from the underlying data dict. This is useful because it
allows us to restore those sections to exactly their previous place. In
contrast, if we delete and add them again, they may end up in a
different position, because the order of the underlying dict can change.

At the moment, the feature is not used anywhere in skops. I plan,
however, to use it for a model card creation space, where it would be
quite useful to have.