Add skops space creator app
Description

This is the code for a streamlit app whose goal it is to let users
easily build HF spaces for sklearn models. In particular, it offers a
GUI to interface with our skops.card.Card class, giving users a wysiwyg
experience.

The app is currently hosted on my personal account:

https://huggingface.co/spaces/BenjaminB/skops-model-card-creator2

We want to move it to skops and update it automatically.

Comments

I have adopted the CI to create this space using skops-ci. If the space
itself is faulty, the CI job will still finish successfully, but at
least one can check it and do some manual testing.

Adding automatic tests will prove quite difficult, because, AFAICT,
streamlit doesn't have anything like a testing client. Automatic testing
would require some heavy lifting from the likes of Selenium or
Playwright.

It is possible to write tests for the parts of the code that are
streamlit-agnostic, e.g. in tasks.py. However, that's the least brittle
part of the code, so the additional value is low.

Right now, there is no job that automatically builds this space and
pushes it to the sklearn orga. I assume for that, we should create a
workflow similar to clean-skops-user.yml that runs regularly and that
uses a secret token so that random people cannot override the app.

I had some strange issues with pre-commit and mypy. I don't want mypy to
check the streamlit app -- it is a bit hacky in some places, which
results in mypy errors. Therefore, I added an exclusion rule to
pyproject.toml, which is indeed respected when I run mypy manually.
However, pre-commit will stil check the repo, even though it's told to
use the pyproject.toml. Thus I had to exclude the space code explicitly
in the pre-commit-config.yaml. If anyone knows what the issue is, please
let me know.