Fix bug, extend tests
The bug was that the attrs of the operator were stored as is, instead of
storing their skops state. This worked in this instance because the
attrs were just a list of str, i.e. primitive types. But still we should
always store to (and restore from) the skops state.

Another unit test was added that tests that operators are not trusted by
default.